refactor(backend): give screen frames their own database module
`database/conversations.py` had grown from 1926 to 2119 lines, tripping the product-file
line-count ratchet. The ratchet offers a declared PR-body exception as the escape hatch,
but +193 lines into an already-oversized file is the kind of growth the ratchet exists to
catch, and the block was one self-contained section.

The screen-frame subcollection helpers move to `database/screen_frames.py`. They are a
conversation subcollection exactly like `photos`, which is why they were written next to
it — the honest reason for the split is the parent file's size, and the new module's
header says so rather than inventing an architectural rationale.

`conversations.py` returns to exactly 1926 lines, its size before this feature.

One non-obvious call site: `tests/unit/test_conversation_events_bounds.py` stubs the whole
`database` package with an explicit submodule allowlist, which could not know about a
module that did not exist when it was written. Missing it turns a hermetic test into one
that reaches for real Firestore and hangs for minutes instead of failing fast.

Verified: 111 screen-frame tests pass in ~1s (the runtime matters — a hermetic suite that
suddenly takes minutes is reaching the network). `import main` clean. Conversation and
photo suites run file-by-file, with three pre-existing failures reproduced identically on
the unmodified checkout, and a collection segfault likewise reproduced there — neither is
caused by this change.
