+++ backend/tests/unit/test_screen_frame_store.py
@@ -14,13 +14,13 @@ CONVERSATION_ID = "conv-1"
 
 class TestDeleteConversationScreenFramesClosesTheLoop:
     def test_deletes_gcs_blobs_for_every_frame_then_the_firestore_docs(self, monkeypatch):
-        fake_conversations_db = MagicMock()
-        fake_conversations_db.get_conversation_screen_frames.return_value = [
+        fake_screen_frames_db = MagicMock()
+        fake_screen_frames_db.get_conversation_screen_frames.return_value = [
             {"id": "frame-a"},
             {"id": "frame-b"},
         ]
-        fake_conversations_db.delete_conversation_screen_frame_docs.return_value = 2
-        monkeypatch.setattr(store_mod, "conversations_db", fake_conversations_db)
+        fake_screen_frames_db.delete_conversation_screen_frame_docs.return_value = 2
+        monkeypatch.setattr(store_mod, "screen_frames_db", fake_screen_frames_db)
 
         fake_storage = MagicMock()
         monkeypatch.setattr(store_mod, "storage", fake_storage)
@@ -31,13 +31,13 @@ class TestDeleteConversationScreenFramesClosesTheLoop:
         assert fake_storage.delete_screen_frame_blobs.call_count == 2
         fake_storage.delete_screen_frame_blobs.assert_any_call(UID, CONVERSATION_ID, "frame-a")
         fake_storage.delete_screen_frame_blobs.assert_any_call(UID, CONVERSATION_ID, "frame-b")
-        fake_conversations_db.delete_conversation_screen_frame_docs.assert_called_once_with(UID, CONVERSATION_ID)
+        fake_screen_frames_db.delete_conversation_screen_frame_docs.assert_called_once_with(UID, CONVERSATION_ID)
 
     def test_no_frames_is_a_clean_no_op(self, monkeypatch):
-        fake_conversations_db = MagicMock()
-        fake_conversations_db.get_conversation_screen_frames.return_value = []
-        fake_conversations_db.delete_conversation_screen_frame_docs.return_value = 0
-        monkeypatch.setattr(store_mod, "conversations_db", fake_conversations_db)
+        fake_screen_frames_db = MagicMock()
+        fake_screen_frames_db.get_conversation_screen_frames.return_value = []
+        fake_screen_frames_db.delete_conversation_screen_frame_docs.return_value = 0
+        monkeypatch.setattr(store_mod, "screen_frames_db", fake_screen_frames_db)
 
         fake_storage = MagicMock()
         monkeypatch.setattr(store_mod, "storage", fake_storage)
@@ -48,9 +48,9 @@ class TestDeleteConversationScreenFramesClosesTheLoop:
 
 class TestDeleteSingleScreenFrame:
     def test_deletes_firestore_doc_and_gcs_blobs(self, monkeypatch):
-        fake_conversations_db = MagicMock()
-        fake_conversations_db.delete_conversation_screen_frame_doc.return_value = True
-        monkeypatch.setattr(store_mod, "conversations_db", fake_conversations_db)
+        fake_screen_frames_db = MagicMock()
+        fake_screen_frames_db.delete_conversation_screen_frame_doc.return_value = True
+        monkeypatch.setattr(store_mod, "screen_frames_db", fake_screen_frames_db)
 
         fake_storage = MagicMock()
         monkeypatch.setattr(store_mod, "storage", fake_storage)
@@ -58,15 +58,15 @@ class TestDeleteSingleScreenFrame:
         existed = store_mod.delete_screen_frame(UID, CONVERSATION_ID, "frame-a")
 
         assert existed is True
-        fake_conversations_db.delete_conversation_screen_frame_doc.assert_called_once_with(
+        fake_screen_frames_db.delete_conversation_screen_frame_doc.assert_called_once_with(
             UID, CONVERSATION_ID, "frame-a"
         )
         fake_storage.delete_screen_frame_blobs.assert_called_once_with(UID, CONVERSATION_ID, "frame-a")
 
     def test_gcs_delete_still_runs_even_if_firestore_doc_was_already_gone(self, monkeypatch):
-        fake_conversations_db = MagicMock()
-        fake_conversations_db.delete_conversation_screen_frame_doc.return_value = False
-        monkeypatch.setattr(store_mod, "conversations_db", fake_conversations_db)
+        fake_screen_frames_db = MagicMock()
+        fake_screen_frames_db.delete_conversation_screen_frame_doc.return_value = False
+        monkeypatch.setattr(store_mod, "screen_frames_db", fake_screen_frames_db)
 
         fake_storage = MagicMock()
         monkeypatch.setattr(store_mod, "storage", fake_storage)
