+++ backend/utils/screen_frames/store.py
@@ -1,10 +1,10 @@
 """Composite (Firestore + GCS + cache) operations for screen frames.
 
-The Firestore-only CRUD lives in database/conversations.py, mirroring the
+The Firestore-only CRUD lives in database/screen_frames.py, mirroring the
 photos subcollection exactly (contract §8). This module is the layer above
 it that also touches GCS — the same split as delete_conversation_audio_files
 (GCS, utils/other/storage.py) being called alongside
-conversations_db.delete_conversation (Firestore-only) from the router,
+database.conversations.delete_conversation (Firestore-only) from the router,
 rather than folding the GCS call into the DB layer.
 """
 
@@ -13,7 +13,7 @@ from __future__ import annotations
 import logging
 from typing import Any, Dict, List
 
-import database.conversations as conversations_db
+import database.screen_frames as screen_frames_db
 import utils.other.storage as storage
 
 logger = logging.getLogger(__name__)
@@ -29,7 +29,7 @@ def persist_screen_frame_docs(uid: str, conversation_id: str, frame_docs: List[D
     """
     if not frame_docs:
         return True
-    return conversations_db.store_conversation_screen_frames(uid, conversation_id, frame_docs)
+    return screen_frames_db.store_conversation_screen_frames(uid, conversation_id, frame_docs)
 
 
 def delete_screen_frame(uid: str, conversation_id: str, frame_id: str) -> bool:
@@ -38,7 +38,7 @@ def delete_screen_frame(uid: str, conversation_id: str, frame_id: str) -> bool:
     unconditionally either way — contract §8: a delete that leaves bytes in
     the bucket is a bug, not a partial success.
     """
-    existed = conversations_db.delete_conversation_screen_frame_doc(uid, conversation_id, frame_id)
+    existed = screen_frames_db.delete_conversation_screen_frame_doc(uid, conversation_id, frame_id)
     storage.delete_screen_frame_blobs(uid, conversation_id, frame_id)
     return existed
 
@@ -51,9 +51,9 @@ def delete_conversation_screen_frames(uid: str, conversation_id: str) -> int:
     §8). It must run before the parent conversation doc is deleted —
     Firestore does not cascade subcollection deletes.
     """
-    frames = conversations_db.get_conversation_screen_frames(uid, conversation_id)
+    frames = screen_frames_db.get_conversation_screen_frames(uid, conversation_id)
     for frame in frames:
         frame_id = frame.get('id')
         if frame_id:
             storage.delete_screen_frame_blobs(uid, conversation_id, frame_id)
-    return conversations_db.delete_conversation_screen_frame_docs(uid, conversation_id)
+    return screen_frames_db.delete_conversation_screen_frame_docs(uid, conversation_id)
